refactor: separated reducer function
